test(reclassify): move importlib-contract test out of the no-growth file
Codex review on #1002 (P1): appending to tests/test_reclassify.py pushed it
from 1781 to 1828 lines, past its architecture/debt.yaml no_growth baseline
-- scripts/check_architecture.py's debt-no-growth check fails outright.
Moved TestNoImportlibWorkaround to a new, properly-scoped
tests/test_reclassify_leaf_import_contract.py instead of raising the
baseline, per AGENTS.md's "Files that are large — edit carefully" section.
tests/test_reclassify.py is back to exactly its 1781-line baseline.
