+++ tests/test_reclassify.py
@@ -1779,50 +1779,3 @@ reclassify:
     validator.validate(report["policy_reclassify"][0]["to"])
     with pytest.raises(jsonschema.exceptions.ValidationError):
         validator.validate("risk")  # the raw YAML spelling, not the schema's
-
-
-class TestNoImportlibWorkaround:
-    """ADR-063 D10 (implementation plan Phase 9): the cycle
-    ``policy_file -> reclassify -> suppression -> checker_types ->
-    policy_file`` that motivated ``_suppression_cls()``'s runtime
-    ``importlib.import_module`` resolution no longer exists — ``reclassify.py``
-    imports ``abicheck.policy.selectors`` (a leaf module with zero dependency
-    on ``suppression.py``/``checker_types.py``/``policy_file.py``/
-    ``finding_identity.py``) statically instead. Confirmed to fail against
-    the pre-Phase-9 code, which had exactly one such call
-    (``_suppression_cls()``, per that module's own docstring)."""
-
-    def test_reclassify_module_contains_no_importlib_import_module_call(self) -> None:
-        import ast
-        import inspect
-
-        import abicheck.reclassify as reclassify_module
-
-        source = inspect.getsource(reclassify_module)
-        tree = ast.parse(source)
-        calls = [
-            node
-            for node in ast.walk(tree)
-            if isinstance(node, ast.Call)
-            and isinstance(node.func, ast.Attribute)
-            and node.func.attr == "import_module"
-        ]
-        assert not calls, (
-            "reclassify.py must not resolve Suppression (or anything else) "
-            "via importlib.import_module -- the import cycle that workaround "
-            "existed for no longer exists (ADR-063 D10)"
-        )
-
-    def test_reclassify_module_does_not_import_importlib(self) -> None:
-        import ast
-        import inspect
-
-        import abicheck.reclassify as reclassify_module
-
-        source = inspect.getsource(reclassify_module)
-        tree = ast.parse(source)
-        imports_importlib = any(
-            isinstance(node, ast.Import) and any(a.name == "importlib" for a in node.names)
-            for node in ast.walk(tree)
-        )
-        assert not imports_importlib
